+++ core/src/main/java/com/crowdar/api/rest/CommonSteps.java
@@ -51,29 +51,33 @@ public class CommonSteps {
         Assert.assertEquals(actualJsonResponse.length, 0);
     }
 
+    @Then("se obtuvo el texto (.*) como response")
+    @And("text (.*) was obtained in response")
+    public void validateTextResponse(String response) {
+        Response actualResponse = APIManager.getLastResponse();
+        Assert.assertEquals(actualResponse.getResponse().toString(), response);
+    }
+
     @Then("se obtuvo el response esperado en ([^ ]*) con el ([^ ]*)")
     @And("expected response is obtained in '([^']*)' with '([^']*)'")
-    public void iWillGetTheProperResponse(String entity, String jsonName) throws IOException, ClassNotFoundException, NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
+    public void iWillGetTheProperResponse(String entity, String jsonName) throws IOException, ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InstantiationException {
         setInjectorParameters(jsonName);
-        Class service = getServiceClass(entity);
-        service.getMethod("validateFields", objectClass, objectClass).invoke(service.newInstance(), expectedJsonResponse, actualJsonResponse);
+        invokeValidateMethod(entity, objectClass, objectClass, expectedJsonResponse, actualJsonResponse);
     }
 
     @Then("se obtuvo el response esperado en ([^ ]*) con el ([^ ]*) y sus parametros ([^ ]*)")
     @And("expected response is obtained in '([^']*)' with '([^ ]*)' and the parameters '([^']*)'")
-    public void iWillGetTheProperResponseWithParameters(String entity, String jsonName, String inputParameters) throws IOException, ClassNotFoundException, NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
+    public void iWillGetTheProperResponseWithParameters(String entity, String jsonName, String inputParameters) throws IOException, ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InstantiationException {
         setInjectorParameters(jsonName);
         Map<String, String> parameters = MapUtils.splitIntoMap(inputParameters, ",", ":");
-        Class service = getServiceClass(entity);
-        service.getMethod("validateFields", objectClass, objectClass, Map.class).invoke(service.newInstance(), expectedJsonResponse, actualJsonResponse, parameters);
+        invokeValidateMethod(entity, parameters, objectClass, expectedJsonResponse);
     }
 
     @Then("se obtuvo el response esperado en ([^ ]*) modificando el ([^ ]*)")
     @And("expected response is obtained in '([^']*)' with '([^']*)' modifying the '([^']*)'")
     public void iWillGetTheProperResponseModified(String entity, String expectedJsonName) throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, IOException, ClassNotFoundException, InstantiationException {
         setInjectorParameters(null);
-        Class service = getServiceClass(entity);
-        service.getMethod("validateFields", String.class, objectClass).invoke(service.newInstance(), "response/".concat(expectedJsonName), actualJsonResponse);
+        invokeValidateMethod(entity, String.class, objectClass, "response/".concat(expectedJsonName), actualJsonResponse);
     }
 
     @Then("se obtuvo el response esperado en ([^ ]*) modificando el ([^ ]*) y sus parametros ([^ ]*)")
@@ -81,25 +85,56 @@ public class CommonSteps {
     public void iWillGetTheProperResponseModifiedWithParameters(String entity, String expectedJsonName, String inputParameters) throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, IOException, ClassNotFoundException, InstantiationException {
         setInjectorParameters(null);
         Map<String, String> parameters = MapUtils.splitIntoMap(inputParameters, ",", ":");
-        Class service = getServiceClass(entity);
-        service.getMethod("validateFields", String.class, objectClass, Map.class).invoke(service.newInstance(), "response/".concat(expectedJsonName), actualJsonResponse, parameters);
+        invokeValidateMethod(entity, parameters, String.class, "response/".concat(expectedJsonName));
     }
 
     @Then("se obtuvo el response esperado en ([^ ]*)")
     @And("expected response is obtained in '([^']*)'")
     public void iWillGetTheProperResponseWithObject(String entity) throws IOException, ClassNotFoundException, NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
         setInjectorParameters(null);
-        Class service = getServiceClass(entity);
-        service.getMethod("validateFields", objectClass).invoke(service.newInstance(), actualJsonResponse);
+        invokeValidateMethod(entity, "validateFields");
     }
 
     @Then("se obtuvo el response esperado en ([^ ]*) y sus parametros ([^ ]*)")
-    @And("expected response is obtained in: '([^']*)' and the parameters '([^']*)'")
+    @And("expected response is obtained in '([^']*)' and the parameters '([^']*)'")
     public void iWillGetTheProperResponseWithObjectAndParameters(String entity, String inputParameters) throws NoSuchMethodException, InvocationTargetException, IllegalAccessException, IOException, ClassNotFoundException, InstantiationException {
         setInjectorParameters(null);
         Map<String, String> parameters = MapUtils.splitIntoMap(inputParameters, ",", ":");
-        Class service = getServiceClass(entity);
-        service.getMethod("validateFields", objectClass, Map.class).invoke(service.newInstance(), actualJsonResponse, parameters);
+        invokeValidateMethod(entity, objectClass, Map.class, actualJsonResponse, parameters);
+    }
+
+    @Then("se obtuvo el response esperado en ([^ ]*) con el metodo ([^ ]*)")
+    @And("expected response is obtained in '([^']*)' with the method '([^']*)'")
+    public void iWillGetTheProperResponseWithObject(String entity, String method) throws IOException, ClassNotFoundException, NoSuchMethodException, InvocationTargetException, IllegalAccessException, InstantiationException {
+        this.setInjectorParameters(null);
+        invokeValidateMethod(entity, method);
+    }
+
+    private void invokeValidateMethod(String entity, String method) throws ClassNotFoundException, IllegalAccessException, NoSuchMethodException, InstantiationException {
+        try {
+            Class service = this.getServiceClass(entity);
+            service.getMethod(method, this.objectClass).invoke(service.newInstance(), this.actualJsonResponse);
+        } catch (InvocationTargetException e) {
+            Assert.fail(e.getCause().toString());
+        }
+    }
+
+    private void invokeValidateMethod(String entity, Class objectClass, Class objectClass2, Object expectedJsonResponse, Object actualJsonResponse) throws ClassNotFoundException, IllegalAccessException, NoSuchMethodException, InstantiationException {
+        try {
+            Class service = getServiceClass(entity);
+            service.getMethod("validateFields", objectClass, objectClass2).invoke(service.newInstance(), expectedJsonResponse, actualJsonResponse);
+        } catch (InvocationTargetException e) {
+            Assert.fail(e.getCause().toString());
+        }
+    }
+
+    private void invokeValidateMethod(String entity, Map<String, String> parameters, Class objectClass, Object expectedJsonResponse) throws ClassNotFoundException, IllegalAccessException, NoSuchMethodException, InstantiationException {
+        try {
+            Class service = getServiceClass(entity);
+            service.getMethod("validateFields", objectClass, objectClass, Map.class).invoke(service.newInstance(), expectedJsonResponse, actualJsonResponse, parameters);
+        } catch (InvocationTargetException e) {
+            Assert.fail(e.getCause().toString());
+        }
     }
 
     private Class getServiceClass(String entity) throws ClassNotFoundException, IllegalAccessException, InvocationTargetException, NoSuchMethodException {