+++ core/src/main/java/com/crowdar/bdd/cukes/hooks/BasicHook.java
@@ -2,8 +2,6 @@ package com.crowdar.bdd.cukes.hooks;
 
 import java.io.IOException;
 
-import org.apache.log4j.Logger;
-
 import com.crowdar.core.Injector;
 import com.crowdar.driver.DriverManager;
 
@@ -12,12 +10,13 @@ import io.cucumber.java.After;
 import io.cucumber.java.AfterStep;
 import io.cucumber.java.Before;
 import io.cucumber.java.BeforeStep;
+import org.apache.log4j.Logger;
 
 public class BasicHook {
 
 	@Before()
 	public void beforeScenario(Scenario scenario) throws IOException{
-		Logger.getRootLogger().info("------ Starting -----" + scenario.getName() + "-----");
+		Logger.getLogger(this.getClass()).info("------ Starting -----" + scenario.getName() + "-----");
 	}
 	
 	@BeforeStep
@@ -32,7 +31,7 @@ public class BasicHook {
 	
 	@After()
 	public void afterScenario(Scenario scenario) throws IllegalAccessException, NoSuchFieldException {
-		Logger.getRootLogger().info("------ Ending -----" + scenario.getName() + "-----");
+		Logger.getLogger(this.getClass()).info("------ Ending -----" + scenario.getName() + "-----");
 		DriverManager.dismissCurrentDriver();
 		Injector.cleanThreadCache();
 	}