+++ core/src/main/java/com/crowdar/core/JsonUtils.java
@@ -1,21 +1,22 @@
 package com.crowdar.core;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.List;
-
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.type.TypeFactory;
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.TagType;
 import com.github.jknack.handlebars.Template;
+import org.apache.log4j.Logger;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.List;
 
 public class JsonUtils {
 
@@ -32,8 +33,8 @@ public class JsonUtils {
         try {
             TypeFactory typeFactory = getMapper().getTypeFactory();
             return (T) mapper.readValue(json, typeFactory.constructType(type));
-        } catch (Exception e) {
-            e.printStackTrace();
+        } catch (IOException e) {
+            Logger.getLogger(JsonUtils.class).error(e.getMessage());
         }
         return null;
     }
@@ -42,8 +43,8 @@ public class JsonUtils {
         String jsonResult = null;
         try {
             jsonResult = getMapper().writeValueAsString(json);
-        } catch (Exception e) {
-            e.printStackTrace();
+        } catch (IOException e) {
+            Logger.getLogger(JsonUtils.class).error(e.getMessage());
         }
         return jsonResult;
     }
@@ -73,12 +74,12 @@ public class JsonUtils {
      * @param fileName
      * @return String (json)
      */
-    public static String getJSONFromFile(String fileName) {
+    public static String getJSONFromFile(String fileName) throws IOException {
         String path = System.getProperty("user.dir").concat(File.separator).concat("src").concat(File.separator).concat("test").concat(File.separator).concat("resources").concat(File.separator).concat("jsons").concat(File.separator).concat(fileName).concat(".json");
         return getJSONFromPath(path);
     }
 
-    public static String getJSONFromPath(String path) {
+    public static String getJSONFromPath(String path) throws IOException {
         return getJSON(Paths.get(path));
     }
 
@@ -100,17 +101,11 @@ public class JsonUtils {
      * @param file
      * @return String (json)
      */
-    public static String getJSON(Path file) {
+    public static String getJSON(Path file) throws IOException {
         ObjectMapper mapper = new ObjectMapper();
-        String json = null;
-        try {
-            FileInputStream fis = new FileInputStream(file.toFile());
-            JsonNode rootNode = mapper.readTree(fis);
-            json = rootNode.toString();
-        } catch (IOException e) {
-            System.out.println("JSON was not found " + file + " " + e.getMessage());
-        }
-        return json;
+        FileInputStream fis = new FileInputStream(file.toFile());
+        JsonNode rootNode = mapper.readTree(fis);
+        return rootNode.toString();
     }
 
     public static String prettyJsonToCompact(String prettyJson) {