+++ core/src/main/java/com/crowdar/core/PDFManager.java
@@ -1,11 +1,15 @@
 package com.crowdar.core;
 
-import java.io.IOException;
-
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.text.PDFTextStripper;
+import org.testng.Assert;
+
+import java.io.IOException;
+import java.util.List;
 
 public class PDFManager {
+
+    private static final String TEXT_NOT_PRESENT_ERROR_MESSAGE = "Text %s is not present in PDF.";
 
     public static String getPdfContent(String fileName) throws IOException {
         String text;
@@ -26,5 +30,12 @@ public class PDFManager {
 
     public static boolean isTextPresentInPDF(String fileName, String expectedText) throws IOException {
         return getPdfContent(fileName).contains(expectedText);
+    }
+
+    public static void isTextPresentInPDF(String fileName, List<String> expectedsText) throws IOException {
+        String pdfContent = getPdfContent(fileName);
+        for (String expectedText : expectedsText) {
+            Assert.assertTrue(pdfContent.contains(expectedText), String.format(TEXT_NOT_PRESENT_ERROR_MESSAGE, expectedText));
+        }
     }
 }